Backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.10.6
ollama==0.4.7
pinecone==3.2.2
sentence-transformers==3.4.1
python-dotenv==1.0.1